+++ src/blitter.rs
@@ -4,7 +4,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-use tiny_skia_path::ScreenIntRect;
+use crate::geom::ScreenIntRect;
 
 use crate::alpha_runs::AlphaRun;
 use crate::color::AlphaU8;