+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Removed
+- `tiny_skia_path::ScreenIntRect`. It become private.
 
 ### Removed
 - `ColorU8::get` and `PremultipliedColorU8::get`. Use the getters instead.